[analytics] move test helpers to new test folder
